requirements.txt
peewee
requests
pre-commit
flake8
hibp-zetsumeishi
